res不会接收到负数，改用int类型

+++ src/heaptimer/heaptimer.cpp
@@ -126,7 +126,7 @@ void HeapTimer::clear() {
 
 int HeapTimer::GetNextTick() {
     tick();
-    size_t res = -1;
+    int res = -1;
     if (!heap_.empty()) {
         res = std::chrono::duration_cast<MS>(heap_.front().expires - Clock::now()).count();
         if (res < 0){ res = 0; }
